AI-Agent/requirements.txt
openai>=1.10.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.10
langchain-community>=0.0.10
python-dotenv
pydantic>=2.0.0
python-magic
requests
wikipedia